+++ src/telstra/mobile/__init__.py
@@ -1,5 +1,6 @@
 from telstra.mobile.account import Prepaid, autodetect_account
 from telstra.mobile.modem import autodetect_modem
+from telstra.mobile.web import TelstraWebApi
 
 
 def debug():